move spf into pyspf package
