+++ pyspf/cid2spf.py
@@ -1,4 +1,4 @@
-#!/usr/bin/python2.4
+#!/usr/bin/python
 
 # Convert a MS Caller-ID entry (XML) to a SPF entry
 # This should be obsolete now.  It is included for historical interest.
